Проводит базовую модуляцию
